% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell2col.r
\name{cell2col}
\alias{cell2col}
\title{Convert Cell Type Names to Colors}
\usage{
cell2col(x)
}
\arguments{
\item{x}{a vector of cell type names.}
}
\value{
a vector of colors for each element in x
}
\description{
Converts cell type names to colors.
Currently the color palette is hard-coded within the function.
Black color is returned for undefined an missing cell types.
}
\examples{
cells <- c("blood", "pbmc", "bcell", "bn", "neu", "mix", "mono")
cell2col(cells)

}
\author{
Karolis KonceviÄius
}
