% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osclines.r
\name{oscline}
\alias{oscline}
\title{Add an oscillation curve}
\usage{
oscline(x0, x1, osc, ...)
}
\arguments{
\item{x0}{a numeric value specifying the starting x-coordinate.}

\item{x1}{a numeric value specifying the ending x-coordinate.}

\item{osc}{a named list or vector containing oscillation parameters:
`mesor`, `acrophase`, `amplitude`, and `period`.}

\item{...}{additional parameters passed to \code{\link[graphics]{lines}}.}
}
\description{
Adds a cosine-based oscillation curve to an existing plot.
}
\details{
The curve is based on 1000 points placed between \code{x0} and \code{x1}.

Requried oscillation parameters:\cr
1. mesor - mean value of the curve.\cr
2. acrophase - timing of the oscillation peak (maximum).\cr
3. amplitude - half the difference between peak and trough.\cr
4. period - length of one complete cycle.
}
\examples{
plot.new()
plot.window(xlim = c(0,24), ylim = c(0,10))
grid()
osc_params <- list(mesor = 5, amplitude = 2, acrophase = 6, period = 24)
oscline(0, 24, osc_params, col = "blue", lwd = 2)

}
\author{
Karolis KonceviÄius
}
