% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redoRDS.r
\name{redoRDS}
\alias{redoRDS}
\title{Rebuild and Read RDS files}
\usage{
redoRDS(file, ...)
}
\arguments{
\item{file}{name of the file where the R object is saved.}

\item{...}{optional arguments passed to readRDS.}
}
\value{
an R object.
}
\description{
Uses "redo" build system to check if the specified .rds file is out of date,
rebuilds it if necessary and then reads the file.
}
\details{
Technically the function first invokes a "redo-ifchange" system command
and then reads the specified .rds file using \code{readRDS()}.
For an example of "redo" implementation see \href{http://www.goredo.cypherpunks.su}.
}
\author{
Karolis Koncevičius
}
