% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex2col.r
\name{sex2col}
\alias{sex2col}
\title{Convert Sex Labels to Colors}
\usage{
sex2col(x)
}
\arguments{
\item{x}{a vector of sex labels.}
}
\value{
a vector of colors for each element in x
}
\description{
Converts sex labels to colors.
Currently the color palette is hard-coded within the function.
Gray color is used for undefined and missing labels.
}
\examples{
sex <- c("F", "M", "Female", "Male", "unknown")
sex2col(sex)

}
\author{
Karolis KonceviÄius
}
