% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iac.r
\name{iac}
\alias{iac}
\title{Inter-Array Correlation}
\usage{
iac(x, groups, sdcutoff = -3, niter = 3)
}
\arguments{
\item{x}{a matrix where features are represented by rows and observations by columns.}

\item{groups}{an optional vector of group membership for each column of \code{x}.}

\item{sdcutoff}{standard deviation cutoff used for outlier detection (default = -3).}

\item{niter}{number of iterations (default = 3).}
}
\value{
a matrix of IAC distances with iterations listed in rows.
}
\description{
Inter-Array Correlation (IAC) method for outlier detection.
The idea behind IAC is to measure the average correlation between each
observation and the rest of the dataset and then mark the observations with
low average correlation as potential outliers.
}
\details{
Technically the function first calculates the average correlations for each
available observation and then transforms those correlations into z-scores.
Only rows with no missing values are used for estimating the correlation
coefficients. For the next iteration all the z-scores below the specified
SD threshold will not be included in the calculations of mean and standard
deviation of average correlation coefficients. However, the actual average
correlation for those observations will still be transformed into a z-score,
reflecting the distance (in standard deviations) from the rest of the group.

When `groups` argument is provided the IAC procedure is performed for each
group separately. When no observations fall below the specified SD threshold
the z-score values returned by IAC will be the same as in returned by the
previous iteration.
}
\examples{
# a toy dataset
dat <- data.matrix(t(iris[,-5]))

# 10 iterations with sd cutoff of -2
res <- iac(dat, sdcutoff = -2, niter = 10)

# only the iterations that found additional outliers
unique(res)

}
\author{
Karolis KonceviÄius
}
